requirements.txt
robotframework
robotframework-requests
rpaframework-pdf
jsonschema
weasyprint